requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.1.0
blosc2==2.2.9
chex==0.1.84
cloudpickle==3.0.0
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distrax==0.1.4
dm-tree==0.1.8
executing==2.0.1
fastjsonschema==2.19.0
-e git+https://github.com/google-deepmind/ferminet.git@b02ea6990b040c2fa87b2e27a1b094254f7bd158#egg=ferminet
folx @ git+https://github.com/microsoft/folx@b3347950d4358022b906a51406aa5cfea213796a
fonttools==4.44.0
gast==0.5.4
h5py==3.10.0
immutabledict==3.0.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
jax==0.4.20
jaxlib==0.4.20+cuda12.cudnn89
jaxtyping==0.2.28
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kfac-jax @ git+https://github.com/deepmind/kfac-jax@2cf83d204683afb3eb668e2507f1e38edc38ad1a
kiwisolver==1.4.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-collections==0.1.1
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.8
numexpr==2.8.7
numpy==1.26.1
nvidia-cublas-cu12==12.3.2.9
nvidia-cuda-cupti-cu12==12.3.52
nvidia-cuda-nvcc-cu12==12.3.52
nvidia-cuda-nvrtc-cu12==12.3.52
nvidia-cuda-runtime-cu12==12.3.52
nvidia-cudnn-cu12==8.9.6.50
nvidia-cufft-cu12==11.0.11.19
nvidia-cusolver-cu12==11.5.3.52
nvidia-cusparse-cu12==12.1.3.153
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.52
opt-einsum==3.3.0
optax==0.1.7
packaging==23.2
pandas==2.1.2
parameterized==0.9.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.5.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyblock==0.6
pydantic==1.10.7
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
pyscf==2.4.0
pytest==8.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.32.0
rpds-py==0.13.2
ruamel.yaml==0.18.4
ruamel.yaml.clib==0.2.8
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
tables==3.9.1
tenacity==8.2.3
tensorflow-probability==0.22.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.9
